Comments added, score text added, added enemy method
Commented all 4 activities
Set up a score text for the game at the moment the text is not appearing
seems to be a problem with background blocking it
Added an enemy method that will spawn enemies.  Is not working at the
moment.
Adjustments will need to be made to show score and fix enemy continual
spawning.
Things to add - Repair collision hits for game, jump mechanic and
possibly some adjustments to the game over activity.